Add ability to specify basic auth headers for influx calls
Some setups do proxying of influxdb calls and use basic auth to do
authentication. Support is added for this via the new BasicUsername and
BasicPassword settings, which enable setting the basic auth username and
password of a request independantly of the username and password passed
on the influx request query parameters.